After adding CRUD features, we added authentication with JWT

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // NOTE get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json({ goals });
 });
 
@@ -20,6 +21,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json({ goal });
 });
@@ -32,6 +34,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found.");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for the user existing and having an acc
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found.");
+  }
+
+  // Check if the user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(400);
+    throw new Error("You are not the owner of this goal.");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -49,6 +65,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found.");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for the user existing and having an acc
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found.");
+  }
+
+  // Check if the user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(400);
+    throw new Error("You are not the owner of this goal.");
   }
 
   await goal.deleteOne({ _id: req.params.id });
